Updated Readme file

+++ lib/services/database.dart
@@ -368,17 +368,29 @@ class DatabaseService {
   Future unlikepost(int initialvalue, String postId, String userEmail) {
 
 
-    Firestore.instance
-        .collection("posts")
-        .document(postId)
-        .updateData({'likes': initialvalue - 1});
+    if(initialvalue < 0){
 
-    Firestore.instance
-        .collection("posts")
-        .document(postId)
-        .collection('likes')
-        .document(userEmail)
-        .delete();
+      initialvalue = 0;
+
+      Firestore.instance
+          .collection("posts")
+          .document(postId)
+          .updateData({'likes': initialvalue});
+    }
+
+    else {
+      Firestore.instance
+          .collection("posts")
+          .document(postId)
+          .updateData({'likes': initialvalue - 1});
+
+      Firestore.instance
+          .collection("posts")
+          .document(postId)
+          .collection('likes')
+          .document(userEmail)
+          .delete();
+    }
   }
 
 //  Future followingUser(int following, String displayNameX, String displayName) async {
